Improvements to the interaction of the Client and SessionCharge models
Client now has a virtual attribute current_rate on Client.  When a Client.new is built, a SessionCharge record is also created with a default value.  If it is an existing record loaded from the database, then the virtual attribute current_rate returns the appropriate value from the relevant SessionCharge record.